requirements.txt
asgiref==3.9.1
attrs==25.3.0
cfgv==3.4.0
distlib==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-gis==1.2.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
filelock==3.18.0
GDAL @ file:///C:/Users/user/Downloads/GDAL-3.11.1-cp313-cp313-win_amd64.whl#sha256=a233e533689df3388ca990f11306dc9e68bf080c34b7460dc4b954500528187e
identify==2.6.12
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
nodeenv==1.9.1
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
ruff==0.12.4
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.32.0
